Added some error checking and messaging

+++ script_multiple2.sh
@@ -5,6 +5,28 @@ export BROWSER=firefox
 #result=`npm test`
 #echo $result
 
+# Test that we got one arg, which should name the site to be tested:
+if [ $# -eq 0 ]; then
+	echo "Please name a site to be tested. Pass the site name as a script argument, like ${0} [sitename]"
+	exit
+fi;
+
+# Test that we have a data file for the site to be tested:
+if [ ! -e "data/$SITE" ]; then
+	echo "We don't have info to test $SITE. The file data/$SITE does not exist."
+	exit
+fi;
+
+if [ ! -f "data/$SITE" ]; then
+        echo "We don't have info to test $SITE. The file data/$SITE isn't a regular file we can read."
+        exit
+fi;
+
+if [ ! -s "data/$SITE" ]; then
+        echo "We don't have info to test $SITE. The file data/$SITE is empty."
+        exit
+fi;
+
 count=0
 rm -rf results/$1
 mkdir results/$1
@@ -20,3 +42,8 @@ do
 	(../../20171020/migrationtest/git_repos/migrationtest/node_modules/webdriverio/bin/wdio wdio.conf.js > results/$1/$count) &
 
 done < data/temp
+
+echo "Started $count tests against $SITE.boisestate.edu. Results will show up in `pwd`/results/$SITE"
+echo "When tests are done, use this command to see if any of them have failed:"
+echo "egrep -l 'failing|ERROR' `pwd`/results/$SITE/*"
+
